fix record handle calculation for 16-bit platforms

+++ trunk/src/sdp.c
@@ -40,7 +40,6 @@
 
 #include "sdp.h"
 
-
 #include <stdio.h>
 #include <string.h>
 
@@ -63,7 +62,7 @@ static void sdp_packet_handler(uint8_t packet_type, uint16_t channel, uint8_t *p
 static linked_list_t sdp_service_records = NULL;
 
 // our handles start after the reserved range
-static uint32_t sdp_next_service_record_handle = maxReservedServiceRecordHandle + 2;
+static uint32_t sdp_next_service_record_handle = ((uint32_t) maxReservedServiceRecordHandle) + 2;
 
 static uint8_t sdp_response_buffer[SDP_RESPONSE_BUFFER_SIZE];
